move sticky header js logic to function
