+++ .eslintrc
@@ -17,6 +17,7 @@
     "jsx-a11y/img-has-alt": ["off"],
     "newline-per-chained-call": ["off"],
     "no-nested-ternary": ["off"],
+    "no-param-reassign": ["off"],
     "no-restricted-globals": ["off"],
     "no-restricted-syntax": ["off"],
     "no-underscore-dangle": ["off"],
